[RestrictedRolePerms] add success message and "any" feature

+++ restrictedroleperms/restrictedroleperms.py
@@ -1,6 +1,8 @@
 from redbot.core import commands, Config
 from redbot.core.utils.chat_formatting import humanize_list
 import discord
+import typing
+from .rrp_converters import ExplicitAll
 
 
 class RestrictedRolePerms(commands.Cog):
@@ -14,8 +16,8 @@ class RestrictedRolePerms(commands.Cog):
         self.bot = bot
         self.config = Config.get_conf(self, identifier=14000605, force_registration=True)
         default_guild = {
-            "mentionable": {"toggle": False, "rules": {}, "message": None},
-            "assignable": {"toggle": False, "rules": {}, "message": None}
+            "mentionable": {"toggle": False, "rules": {}, "message": None, "success": None},
+            "assignable": {"toggle": False, "rules": {}, "message": None, "success": None}
         }
         self.config.register_guild(**default_guild)
 
@@ -26,6 +28,8 @@ class RestrictedRolePerms(commands.Cog):
             found = rules.get(ar)
             if found and role1 in found:
                 perms = True
+            elif found and "any" in found:
+                perms = None
         print(rules.keys(), author_roles)
         return perms
 
@@ -42,7 +46,8 @@ class RestrictedRolePerms(commands.Cog):
         if not rules['toggle']:
             return await ctx.send("The mentionability feature is toggled off for this server.")
 
-        if not await self._has_rule(rules['rules'], [str(r.id) for r in ctx.author.roles], role.id):
+        has_perms = await self._has_rule(rules['rules'], [str(r.id) for r in ctx.author.roles], role.id)
+        if (has_perms is False) or (has_perms is None and role >= ctx.author.top_role):
             if rules['message']:
                 return await ctx.send(rules['message'])
             return await ctx.send(f"Unfortunately, no rules were found allowing you to toggle mentionability for {role.mention}.")
@@ -54,6 +59,9 @@ class RestrictedRolePerms(commands.Cog):
         except discord.HTTPException:
             return await ctx.send("Something went wrong while editing this role.")
 
+        if rules['success']:
+            return await ctx.send(rules['success'][0].replace("{role}", role.name).replace("{role.mention}", role.mention))
+
         return await ctx.send(f"{role.mention} has been set to be mentionable by everyone.")
 
     @_rrp.command(name="denymentions")
@@ -64,7 +72,8 @@ class RestrictedRolePerms(commands.Cog):
         if not rules['toggle']:
             return await ctx.send("The mentionability feature is toggled off for this server.")
 
-        if not await self._has_rule(rules['rules'], [str(r.id) for r in ctx.author.roles], role.id):
+        has_perms = await self._has_rule(rules['rules'], [str(r.id) for r in ctx.author.roles], role.id)
+        if (has_perms is False) or (has_perms is None and role >= ctx.author.top_role):
             if rules['message']:
                 return await ctx.send(rules['message'])
             return await ctx.send(f"Unfortunately, no rules were found allowing you to toggle mentionability for {role.mention}.")
@@ -76,6 +85,9 @@ class RestrictedRolePerms(commands.Cog):
         except discord.HTTPException:
             return await ctx.send("Something went wrong while editing this role.")
 
+        if rules['success']:
+            return await ctx.send(rules['success'][1].replace("{role}", role.name).replace("{role.mention}", role.mention))
+
         return await ctx.send(f"{role.mention} has been set to be not mentionable by everyone.")
 
     @_rrp.command(name="assignrole")
@@ -86,7 +98,8 @@ class RestrictedRolePerms(commands.Cog):
         if not rules['toggle']:
             return await ctx.send("The role assignment feature is toggled off for this server.")
 
-        if not await self._has_rule(rules['rules'], [str(r.id) for r in ctx.author.roles], role.id):
+        has_perms = await self._has_rule(rules['rules'], [str(r.id) for r in ctx.author.roles], role.id)
+        if (has_perms is False) or (has_perms is None and role >= ctx.author.top_role):
             if rules['message']:
                 return await ctx.send(rules['message'])
             return await ctx.send(f"Unfortunately, no rules were found allowing you to assign {role.mention}.")
@@ -98,6 +111,9 @@ class RestrictedRolePerms(commands.Cog):
         except discord.HTTPException:
             return await ctx.send("Something went wrong while assigning this role.")
 
+        if rules['success']:
+            return await ctx.send(rules['success'][0].replace("{role}", role.name).replace("{role.mention}", role.mention).replace("{member}", f"{member.display_name}").replace("{member.mention}", member.mention))
+
         return await ctx.send(f"{role.mention} has been assigned to {member.mention}.")
 
     @_rrp.command(name="removerole")
@@ -108,7 +124,8 @@ class RestrictedRolePerms(commands.Cog):
         if not rules['toggle']:
             return await ctx.send("The role assignment feature is toggled off for this server.")
 
-        if not await self._has_rule(rules['rules'], [str(r.id) for r in ctx.author.roles], role.id):
+        has_perms = await self._has_rule(rules['rules'], [str(r.id) for r in ctx.author.roles], role.id)
+        if (has_perms is False) or (has_perms is None and role >= ctx.author.top_role):
             if rules['message']:
                 return await ctx.send(rules['message'])
             return await ctx.send(f"Unfortunately, no rules were found allowing you to assign {role.mention}.")
@@ -122,6 +139,9 @@ class RestrictedRolePerms(commands.Cog):
             return await ctx.send("I do not have permissions to remove this role!")
         except discord.HTTPException:
             return await ctx.send("Something went wrong while removing this role.")
+
+        if rules['success']:
+            return await ctx.send(rules['success'][1].replace("{role}", role.name).replace("{role.mention}", role.mention).replace("{member}", f"{member.display_name}").replace("{member.mention}", member.mention))
 
         return await ctx.send(f"{role.mention} has been removed from {member.mention}.")
 
@@ -142,7 +162,8 @@ class RestrictedRolePerms(commands.Cog):
 
         mentionable_rules = f"""
         **Toggle:** {mentionable["toggle"]}
-        **Message:** {mentionable['message'] if mentionable['message'] else "Default"}
+        **Error Message:** {mentionable['message'] if mentionable['message'] else "Default"}
+        **Success Message:** {f"{mentionable['success'][0]}, {mentionable['success'][1]}" if mentionable['success'] else "Default"}
         **Rules:** {"None" if not mentionable['rules'] else ""}\n"""
         for r0, r1 in mentionable["rules"].items():
             mentionable_rules += f"{ctx.guild.get_role(int(r0)).name} can toggle mentionability for {humanize_list([ctx.guild.get_role(int(r2)) for r2 in r1])}"
@@ -150,7 +171,8 @@ class RestrictedRolePerms(commands.Cog):
 
         assignable_rules = f"""
         **Toggle:** {assignable["toggle"]}
-        **Message:** {assignable['message'] if assignable['message'] else "Default"}
+        **Error Message:** {assignable['message'] if assignable['message'] else "Default"}
+        **Success Message:** {f"{assignable['success'][0]}, {assignable['success'][1]}" if assignable['success'] else "Default"}
         **Rules:** {"None" if not assignable['rules'] else ""}\n"""
         for r0, r1 in assignable["rules"].items():
             assignable_rules += f"{ctx.guild.get_role(int(r0)).name} can assign {humanize_list([ctx.guild.get_role(int(r2)) for r2 in r1])}"
@@ -163,40 +185,56 @@ class RestrictedRolePerms(commands.Cog):
         """Add a rule to give certain roles restricted permissions."""
 
     @_add_rule.command(name="mentionable")
-    async def _add_mentionable(self, ctx: commands.Context, role_to_give_perms_to: discord.Role, *roles_to_allow_to_be_made_mentionable: discord.Role):
-        """Allow a certain role to make a few other roles mentionable through RRP."""
+    async def _add_mentionable(self, ctx: commands.Context, role_to_give_perms_to: discord.Role, *roles_to_allow_to_be_made_mentionable: typing.Union[discord.Role, ExplicitAll]):
+        """
+        Allow a certain role to make a few other roles mentionable through RRP.
+
+        For `roles_to_allow_to_be_made_mentionable`, either input a list of roles or `all` to allow perms for all roles below the given role.
+        """
 
         # Hierarchy checks
         if role_to_give_perms_to >= ctx.author.top_role and ctx.author != ctx.guild.owner:
             return await ctx.send("The role you want to give perms to is above you in the role hierarchy!")
-        for rm in roles_to_allow_to_be_made_mentionable:
-            if rm >= role_to_give_perms_to:
-                return await ctx.send(f"{rm.mention} is above {role_to_give_perms_to.mention} in the role hierarchy!")
+        if "all" not in roles_to_allow_to_be_made_mentionable:
+            for rm in roles_to_allow_to_be_made_mentionable:
+                if rm >= role_to_give_perms_to:
+                    return await ctx.send(f"{rm.mention} is above {role_to_give_perms_to.mention} in the role hierarchy!")
 
         async with self.config.guild(ctx.guild).mentionable.rules() as rules:
             if rules.get(str(role_to_give_perms_to.id)):
                 return await ctx.send("There is already a rule for that role! Please remove it first using `[p]rrpset removerule`.")
-            rules[str(role_to_give_perms_to.id)] = [r.id for r in roles_to_allow_to_be_made_mentionable]
-
-        return await ctx.send(f"{role_to_give_perms_to.mention} is now allowed to toggle mentionability for {humanize_list([r.mention for r in roles_to_allow_to_be_made_mentionable])}")
+            if "all" not in roles_to_allow_to_be_made_mentionable:
+                rules[str(role_to_give_perms_to.id)] = [r.id for r in roles_to_allow_to_be_made_mentionable]
+                return await ctx.send(f"{role_to_give_perms_to.mention} is now allowed to toggle mentionability for {humanize_list([r.mention for r in roles_to_allow_to_be_made_mentionable])}")
+            else:
+                rules[str(role_to_give_perms_to.id)] = ["all"]
+                return await ctx.send(f"{role_to_give_perms_to.mention} is now allowed to toggle mentionability for all roles below it.")
 
     @_add_rule.command(name="assignable")
-    async def _add_assignable(self, ctx: commands.Context, role_to_give_perms_to: discord.Role, *roles_to_allow_to_be_assigned: discord.Role):
-        """Allow a certain role to assign a few other roles through RRP."""
+    async def _add_assignable(self, ctx: commands.Context, role_to_give_perms_to: discord.Role, *roles_to_allow_to_be_assigned: typing.Union[discord.Role, ExplicitAll]):
+        """
+        Allow a certain role to assign a few other roles through RRP.
+
+        For `roles_to_allow_to_be_assigned`, either input a list of roles or `all` to allow perms for all roles below the given role.
+        """
 
         # Hierarchy checks
         if role_to_give_perms_to >= ctx.author.top_role and ctx.author != ctx.guild.owner:
             return await ctx.send("The role you want to give perms to is above you in the role hierarchy!")
-        for ra in roles_to_allow_to_be_assigned:
-            if ra >= role_to_give_perms_to:
-                return await ctx.send(f"{ra.mention} is above {role_to_give_perms_to.mention} in the role hierarchy!")
+        if "all" not in roles_to_allow_to_be_assigned:
+            for ra in roles_to_allow_to_be_assigned:
+                if ra >= role_to_give_perms_to:
+                    return await ctx.send(f"{ra.mention} is above {role_to_give_perms_to.mention} in the role hierarchy!")
 
         async with self.config.guild(ctx.guild).assignable.rules() as rules:
             if rules.get(str(role_to_give_perms_to.id)):
                 return await ctx.send("There is already a rule for that role! Please remove it first using `[p]rrpset removerule`.")
-            rules[str(role_to_give_perms_to.id)] = [r.id for r in roles_to_allow_to_be_assigned]
-
-        return await ctx.send(f"{role_to_give_perms_to.mention} is now allowed to assign {humanize_list([r.mention for r in roles_to_allow_to_be_assigned])}")
+            if "all" not in roles_to_allow_to_be_assigned:
+                rules[str(role_to_give_perms_to.id)] = [r.id for r in roles_to_allow_to_be_assigned]
+                return await ctx.send(f"{role_to_give_perms_to.mention} is now allowed to assign {humanize_list([r.mention for r in roles_to_allow_to_be_assigned])}")
+            else:
+                rules[str(role_to_give_perms_to.id)] = ["all"]
+                return await ctx.send(f"{role_to_give_perms_to.mention} is now allowed to assign all roles below it.")
 
     @_rrpset.group(name="removerule")
     async def _remove_rule(self, ctx: commands.Context):
@@ -254,4 +292,37 @@ class RestrictedRolePerms(commands.Cog):
     async def _no_rule_message_assignable(self, ctx: commands.Context, *, message: str = None):
         """Customize the role assignment error message (leave empty to reset)."""
         await self.config.guild(ctx.guild).assignable.message.set(message)
+        return await ctx.tick()
+
+    @_rrpset.group(name="successmessage")
+    async def _success_message(self, ctx: commands.Context):
+        """Customize the error message sent when no rules are found giving the user restricted permissions."""
+
+    @_success_message.command(name="mentionable")
+    async def _success_message_mentionable(self, ctx: commands.Context, *, message: str = None):
+        """
+        Customize the mentionability success message (leave empty to reset).
+
+        You can put in `{role}` or `{role.mention}`, which will be replaced with the role that is modified.
+        Please format the message like so: `message for allowmentions//message for denymentions`.
+        """
+        success_messages = message.split("//")
+        if not len(success_messages) > 1:
+            return await ctx.send("Please separate the 2 messages with `//`.")
+        await self.config.guild(ctx.guild).mentionable.success.set((success_messages[0].strip(), success_messages[1].strip()))
+        return await ctx.tick()
+
+    @_success_message.command(name="assignable")
+    async def _success_message_assignable(self, ctx: commands.Context, *, message: str = None):
+        """
+        Customize the role assignment success message (leave empty to reset).
+
+        You can put in `{role}` or `{role.mention}`, which will be replaced with the role that is assigned/removed.
+        Also, you can put in `{member}` or `{member.mention}`, which will be replaced with the member.
+        Please format the message like so: `message for assignrole//message for removerole`.
+        """
+        success_messages = message.split("//")
+        if not len(success_messages) > 1:
+            return await ctx.send("Please separate the 2 messages with `//`.")
+        await self.config.guild(ctx.guild).assignable.success.set((success_messages[0].strip(), success_messages[1].strip()))
         return await ctx.tick()
